+++ src/tablecloth/api.clj
@@ -1,6 +1,7 @@
 (ns tablecloth.api
   (:refer-clojure :exclude [group-by drop concat rand-nth first last shuffle])  
-  (:require [tech.v3.datatype.export-symbols :as exporter]))
+  (:require [tech.v3.datatype.export-symbols :as exporter]
+            [tech.v3.dataset :as ds]))
 
 (exporter/export-symbols tech.v3.datatype
                          clone)
@@ -27,6 +28,7 @@
 
 (exporter/export-symbols tablecloth.api.dataset
                          dataset?
+                         empty-ds?
                          dataset
                          shape
                          info
@@ -126,4 +128,3 @@
 
 (def select (partial select-or-drop select-columns select-rows))
 (def drop (partial select-or-drop drop-columns drop-rows))
-