+++ src/tablecloth/api/unique_by.clj
@@ -1,14 +1,14 @@
 (ns tablecloth.api.unique-by
   (:require [tech.v3.dataset :as ds]
             [tech.v3.datatype.protocols :as dtype-proto]
-            
-            [tablecloth.api.utils :refer [iterable-sequence? column-names]]
-            [tablecloth.api.dataset :refer [dataset]]
-            [tablecloth.api.columns :refer [select-columns]]
-            [tablecloth.api.group-by :refer [grouped? process-group-data ungroup]]
             [tech.v3.dataset.column :as col]
             [tech.v3.datatype :as dtype]
-            [tech.v3.datatype.bitmap :as bitmap]))
+            [tech.v3.datatype.bitmap :as bitmap]
+            
+            [tablecloth.api.utils :refer [iterable-sequence? column-names]]
+            [tablecloth.api.dataset :refer [dataset empty-ds?]]
+            [tablecloth.api.columns :refer [select-columns]]
+            [tablecloth.api.group-by :refer [grouped? process-group-data ungroup]]))
 
 (defn- strategy-first [_ idxs] (clojure.core/first idxs))
 (defn- strategy-last [_ idxs] (clojure.core/last idxs))
@@ -76,6 +76,10 @@
   [ufn ds]
   (if (= 1 (ds/row-count ds)) ds (ufn ds)))
 
+(defn- maybe-empty
+  [ufn ds]
+  (if (empty-ds? ds) ds (ufn ds)))
+
 (defn unique-by
   ([ds] (unique-by ds (ds/column-names ds)))
   ([ds columns-selector] (unique-by ds columns-selector nil))
@@ -84,8 +88,9 @@
                          :as options}]
    (let [selected-keys (column-names ds select-keys)
          ufn (unique-by-fn strategy columns-selector selected-keys options)
-         ufn (if (fn? strategy) ufn (partial maybe-skip-unique ufn))]
+         ufn (partial maybe-empty (if (fn? strategy) ufn (partial maybe-skip-unique ufn)))]
 
      (if (grouped? ds)
        (process-group-data ds ufn parallel?)
        (ufn ds)))))
+